refactor(fast_io): split at_syscalls into per-syscall submodules
