+++ crates/fast_io/src/dir_sandbox/at_syscalls/lstat.rs
@@ -1,0 +1,108 @@
+//! lstat-class SEC-1.f adaptor and the shared single-component-leaf
+//! resolver.
+//!
+//! [`LstatOutcome`] unifies the sandbox-anchored
+//! `fstatat(AT_SYMLINK_NOFOLLOW)` result and the path-based
+//! [`std::fs::symlink_metadata`] fallback behind a `dev`/`ino` surface.
+//! [`single_component_leaf`] is the gate every `*_via_sandbox_or_fallback`
+//! helper in this module uses to decide whether the sandbox fast path is
+//! eligible.
+
+use std::ffi::OsStr;
+use std::io;
+use std::path::Path;
+
+use super::AtMetadata;
+use super::metadata::fstatat_nofollow;
+
+/// Result of [`lstat_via_sandbox_or_fallback`].
+///
+/// The variant indicates which lstat path satisfied the call. Both
+/// variants expose `dev` / `ino` so the hardlink quick-check can
+/// compare inode identity without caring which syscall produced the
+/// numbers.
+#[derive(Debug)]
+pub enum LstatOutcome {
+    /// Sandbox-anchored `fstatat(AT_SYMLINK_NOFOLLOW)` result.
+    At(AtMetadata),
+    /// Path-based [`std::fs::symlink_metadata`] result used when the
+    /// sandbox was unavailable or the relative path was not a single
+    /// component.
+    Std(std::fs::Metadata),
+}
+
+impl LstatOutcome {
+    /// Device id of the entry.
+    #[must_use]
+    pub fn dev(&self) -> u64 {
+        match self {
+            Self::At(meta) => meta.dev(),
+            Self::Std(meta) => std::os::unix::fs::MetadataExt::dev(meta),
+        }
+    }
+
+    /// Inode number of the entry.
+    #[must_use]
+    pub fn ino(&self) -> u64 {
+        match self {
+            Self::At(meta) => meta.ino(),
+            Self::Std(meta) => std::os::unix::fs::MetadataExt::ino(meta),
+        }
+    }
+}
+
+/// Issue `fstatat(AT_SYMLINK_NOFOLLOW)` against `link_path` when the
+/// `sandbox` root is the immediate parent.
+///
+/// SEC-1.f adaptor for callers that already have an absolute path:
+/// - When `sandbox` is `Some`, `link_path` equals
+///   `dest_dir.join(relative_path)`, and `relative_path` has a single
+///   component, the helper resolves the leaf through the sandbox dirfd
+///   so a mid-syscall symlink swap on the leaf cannot redirect the
+///   stat.
+/// - In every other case the helper falls back to
+///   [`std::fs::symlink_metadata`] on `link_path`.
+///
+/// # Errors
+///
+/// Surfaces either the [`fstatat_nofollow`] error or the
+/// [`std::fs::symlink_metadata`] error verbatim, depending on which
+/// path was taken.
+pub fn lstat_via_sandbox_or_fallback(
+    sandbox: Option<&crate::dir_sandbox::DirSandbox>,
+    dest_dir: &Path,
+    relative_path: &Path,
+    link_path: &Path,
+) -> io::Result<LstatOutcome> {
+    if let Some(sandbox) = sandbox
+        && let Some(leaf) = single_component_leaf(dest_dir, relative_path, link_path)
+    {
+        return fstatat_nofollow(sandbox.current_dirfd(), leaf).map(LstatOutcome::At);
+    }
+    std::fs::symlink_metadata(link_path).map(LstatOutcome::Std)
+}
+
+/// Returns the leaf component of `link_path` when `link_path` is
+/// exactly `dest_dir` joined with a single-component `relative_path`.
+///
+/// Multi-component relative paths need a per-directory dirfd stack
+/// (SEC-1.f's follow-up work), so they take the path-based fallback
+/// for now.
+pub(super) fn single_component_leaf<'a>(
+    dest_dir: &Path,
+    relative_path: &'a Path,
+    link_path: &Path,
+) -> Option<&'a OsStr> {
+    let mut comps = relative_path.components();
+    let first = match comps.next()? {
+        std::path::Component::Normal(name) => name,
+        _ => return None,
+    };
+    if comps.next().is_some() {
+        return None;
+    }
+    if dest_dir.join(relative_path) != link_path {
+        return None;
+    }
+    Some(first)
+}
